Fix lve_packages.sh script and cloudlinux config
Modified lve_packages.sh script to include an additional option for
reseller that is needed in the new CloudLinux version or you can't
properly list the packages.
Also in puppet code the CUSTOM_GETPACKAGE_SCRIPT line was appended at
the end of the file but no newline was added in front of the line.

Resolves compatiblity issues with some tools.

Changelog
-