New puppetmaster provision script for RHEL
Added new provisioning script for CentOS 7. This way we can ensure that
puppetmaster is installed and configured correctly. This new setup
will be deployed on new client installations since we don't support
Ubuntu 14.04 anymore, than CentOS 7 must be the new puppetmaster.
The only issue is that we need a custom repository PUIAS enabled.
This is done by the script, but it's a third party repo.

Partially resolves PROD-1931.