Fix vagrant master DNS settings, Atomia installer
After executing vagrant up master the machine did not provision without
first modifiying the Ethernet interface to have 127.0.0.1 and 8.8.8.8 as
DNS servers. Now these two DNS entries are added and chocolatey is able
to downloaded the needed applications and Atomia installer.

The other issues is that Atomia installer was not installed
because it was not available anymore from the old installer.atomia.com
url. Now we fetch Atomia installer from:
 installer.atomia.com/AtomiaInstaller-latest.exe
This allows us to always update the  installer when needed.

In order for this to be applied developers would need to either clone
new vagrant-atomia-repo or update the submodules.

Resolves PROD-2143.

ChangeLog:
* [FIX] - DNS settings on Vagrant now setup correctly.
* [FIX] - Newest Atomia installer is now downloaded and installed.