Fix AD join errors and glusterfs
There was an issue with joining fsagent, glusterfs and apache to a
active directory domain. The issue was with using unscd instead of
relying on nscd. Caching for nscd has been disabled only on fsagent.

Issues with remounting storage were also resolved by using newer version
of puppet module for fstab. Changes are needed to the vagrant-atomia to
be applied on vagrant, because of using older module for fstab. This is
why this then partially resolves this issue for vagrant.

Partially resolves PROD-2144.