Set TestRepository as default for vagrant
By default when Atomia applications are installed via AtomiaCMD
installer they are installed from the TestRepo. After when Atomia
installer is installed on the system default repo is set to
PublicRepository which is not needed for development. In order for the
environment to be ready for developers and QA TestRepository is set by
default for the Atomia installer when provisioned via vagrant.

Resolves PROD-2164.